refactor: Move classes to new keymap package
