+++ app/src/main/aidl/xtr/keymapper/OnKeyEventListener.aidl
@@ -1,7 +1,5 @@
 package xtr.keymapper;
 
-// Declare any non-default types here with import statements
-
 interface OnKeyEventListener {
     void onKeyEvent(String event);
 }
