+++ app/src/main/aidl/xtr/keymapper/keymap/KeymapConfig.aidl
@@ -1,0 +1,3 @@
+package xtr.keymapper.keymap;
+
+parcelable KeymapConfig;
