+++ app/src/main/aidl/xtr/keymapper/keymap/KeymapProfile.aidl
@@ -1,0 +1,3 @@
+package xtr.keymapper.keymap;
+
+parcelable KeymapProfile;
