+++ app/src/main/java/xtr/keymapper/server/IInputInterface.java
@@ -1,8 +1,8 @@
 package xtr.keymapper.server;
 
 import xtr.keymapper.IRemoteServiceCallback;
-import xtr.keymapper.KeymapConfig;
-import xtr.keymapper.KeymapProfile;
+import xtr.keymapper.keymap.KeymapConfig;
+import xtr.keymapper.keymap.KeymapProfile;
 import xtr.keymapper.touchpointer.KeyEventHandler;
 import xtr.keymapper.touchpointer.MouseEventHandler;
 
@@ -16,4 +16,5 @@ public interface IInputInterface {
     IRemoteServiceCallback getCallback();
     void moveCursorX(float x);
     void moveCursorY(float y);
+    void reloadKeymap();
 }
